p2p/node.go: rename Send to Broadcast
This is to reflect the nature of communication.
We're not really considering sending messages at a higher level.

Of course this is what does happen on the lower level, i.e. we pack
messages as bytes and send them over open sockers, however when
considering the communication protocol, we always need to send all
messages to all participants of the P2P network.