rfs/root.go: return file descriptor related to rfsFile in create
Whenever a file was created and written to in one go (e.g. when creating
it for the first time in vim), it would only propagate the CREATE
operation.

This was caused by the fact that we would call the method from the
underlying LoopbackNode which would return its own implementation of
file descriptor handler.

The CREATE operation returns a file descriptor to its caller. Then, the
caller acts upon this file descriptor. The file descriptor that was
returned from our Create function would not be our implementation.

Because of that, our hooks for propagating this change to all
participating nodes would not get executed.

I reimplemented most of the logic from go-fuse code and changed it a bit
so that it returns a file descriptor and an inode object related to
rfsFile.